moved simulate_* to a subfolder
