+++ macrosynergy/management/simulate/simulate_vintage_data.py
